Change invisible hitbox strategy and sizing to prevent a nasty stutter
If buttons were hovered just below the trigger point, the scale-expanding of the invisible hitbox would trigger the button animation to stutter up and down in an ugly manner. This fixes that by using a `display` toggle instead, ensuring the button is going to continue going upwards until the mouse either moves out of the lenient spot or the animation ends.