fix(docker): Move install to last step
------------
The Problems
------------
There are namely two problems which are being fixed in this issue:

1. `pyproject.toml` defines `modmail` package to be included, thus
poetry searches for it but finds it be to be empty as originally
we copied the poetry meta files only before installing and then
copied rest of the package.

2. We are using pip to install poetry, we uses the new installer,
the new installer has a known bug see python-poetry/poetry#3336 but
the issue hasn't shown any acitivty since it has been opened
unfortunately.

------------
The Solution
-----------

1. I have moved `COPY . .` to be ran before poetry install so
poetry can find what the package to be included.

2. The only possible solution to this is to disable the old
installer which was causing this bug for some reason.