fix(exts): Fix all loaded cogs check while resyncing cogs
Originally when you would resync the cogs, it would check
whether `(ext, status)` is present in the extensions which
is a dict. But this would always return `False`, for example

```pycon
a = {"a": "a"}

>>> print(("a", "a") in a)
False
```

A fix to this, as committed is to just check whether `ext`
is present in the extension keys.