chore: remove support for webhooks
Using webhooks to send messages do not add much if any benefit, as
a webhook with the member username and pfp is a bit harder to tell apart
from staff at first glance. This is also the only benefit that using
webhooks give, for the multiple disadventages they give, as now the bot
has to manage a webhook too.

We may support webhooks in the future as an official plugin, but for now
we are focusing on the core functionality of the bot.