feat(db): Add init_db function to initalize the DB
This function is ran before the discord modmail bot is ran so that the
bot is not ran unnecessarily even if the DB is not responding. It
creates a local session which can be used throughout the bot and then
tests if the db is responding.

If the DB is not responding it would log the error with the exception
and then exit, it doesn't raise the exception as the the sqlalchemy
connection errors are usually really long and maybe "spammy", to reduce
the logs down it just logs down the error and then exists silently.