revert(db): Remove Embed model json validation
Since we are getting the embed json content directly via discord.py's
embed and not generating or touching it, we are safe and validation
shouldn't be needed.

I have made an extra commit here, in case we ever need it again in
future we won't need to rewrite it all over again.