plugins: refactor to allow local plugins
local plugins now act mostly like normal plugins.
they do not have any special names, and exist as if they were installed normally

to add a local plugin, put the files in the plugins folder,
in a normal folder, and create a local.toml file in the same format as
plugins.toml for addon repositories.