WIP add script to generate requirements.txt file
Purpose of this is to not require poetry to export a requirements.txt file
poetry is big, and has a lot of dependencies, so anything that doesn't
require poetry is great.

This script is still only partial of exports, and also checks if the
poetry.lock file is up to date before commiting.

Support for python version markers and platform markers still need to
be added to the export.

Extras are also missing from the export.