tests: don't always test the mocks
The mock file is pretty big, and due to that has a lot of tests
The mocks should not need to be edited often, only when the dependency
they are mocking is updated. This is currently never, as it is unmaintained.
This means that we don't need to run the tests for the mocks all the time.

If an upgrade of our dependencies is taken, then the test file should be ran.
However, it does not need to run every time the test suite is ran.
This should lead to a faster total test suite when running the suite.